Fix win64 warning C4267: conversion from 'size_t' to 'int'
git-svn-id: file:///home/jj/hercules.svn/trunk@5171 956126f8-22a0-4046-8f4a-272fa8102e63

+++ sllib.c
@@ -12,6 +12,9 @@
 // $Id$
 //
 // $Log$
+// Revision 1.17  2009/01/23 13:02:48  bernard
+// copyright notice
+//
 // Revision 1.16  2007/06/23 00:04:15  ivan
 // Update copyright notices to include current year (2007)
 //
@@ -918,7 +921,7 @@ sl_vol( SLLABEL *lab,
         char *volser,
         char *owner )
 {
-    int len;
+    size_t len;
 
     /*
     || Initialize
@@ -1071,8 +1074,8 @@ sl_ds1( SLLABEL *lab,
         int blocks )
 {
     int gdg;
-    int len;
-    int ndx;
+    size_t len;
+    size_t ndx;
     char wbuf[ 80 ];
 
     /*
@@ -1324,7 +1327,7 @@ sl_ds2( SLLABEL *lab,
         char *trtch )
 {
     int i;
-    int len;
+    size_t len;
     char wbuf[ 80 ];
 
     /*
@@ -1612,7 +1615,7 @@ sl_usr( SLLABEL *lab,
         int num,
         char *data )
 {
-    int len;
+    size_t len;
 
     /*
     || Initialize
